remove external control info remnants
